Adjust touch-only media query
Samsung phones incorrectly report hover as being available, which leads
to the "hover" media query being active. This is a problem because the
delete buttons are not visible on Samsung phones.

Using `pointer: coarse` avoids this, by also showing these elements on
devices without a mouse. This could negatively impact users who use a
stylus, but I think it is safe to assume that we have more Samsung users
than stylus users.

Relevant blog post:
https://www.ctrl.blog/entry/css-media-hover-samsung.html